Modded Organisation of C_CPP directory
Modded Organization cuz that's wot I do best
