+++ C_CPP/Fundamentals/Static scoping & Dynamic Scoping/readme.md
