+++ C_CPP/OOPs in C++/Inheritance and its types or Function overriding/Hierarchical_Inheritance.md
