+++ C_CPP/OOPs in C++/Virtual Functions, Friend Functions & Friend Classes/Virtual_Function.md
