+++ C_CPP/readme.md
@@ -17,85 +17,102 @@
 
 # Topics in the C++ language &nbsp;&nbsp;&nbsp;&nbsp; ![C](https://img.shields.io/badge/C-black?style=for-the-badge&logo=c&labelColor=black&color=404040) &nbsp;&nbsp; ![C++](https://img.shields.io/badge/CPP-blue?style=for-the-badge&logo=cplusplus&labelColor=006199)
 
-- ### [Introduction of C](Introduction%20of%20C/)
-
-- ### [Programming components, Structure & Keyword](Programming%20components%2C%20Structure%20%26%20Keyword/)
-
-- ### [Operators & Expressions](Operators%20%26%20Expressions/)
-  
-- ### [Input & Output Operations](Input%20%26%20Output%20Operations/)
-
-- ### [Decision Making & Branching Structure](Decision%20Making%20%26%20Branching%20Structure/)
-
-- ### [Looping Control Structure/Statements](Looping%20Control%20Structure%20or%20Statements/)
-
-- ### [Break/Continue](Break%20or%20Continue/) 
-
-- ### [Static scoping & Dynamic Scoping](Static%20scoping%20%26%20Dynamic%20Scoping/)
-
-- ### [Functions & Recursion and its types](Functions%20%26%20Recursion%20and%20its%20types/)
-
-- ### [Arrays](Arrays/)
-
-- ### [Strings and their functions](Strings%20and%20their%20functions/)
-
-- ### [Pointer and its types](Pointer%20and%20its%20types/)
-
-- ### [Procedural Structures](Procedural%20Structures/)
-
-- ### [Dynamic Memory allocation](Dynamic%20Memory%20allocation/)
-
-- ### [Structure padding, Union, Enum, & typedef](Structure%20padding%2C%20Union%2C%20Enum%2C%20%26%20typedef/)
-
-- ### [File Management / Handling](File%20Management%20or%20Handling/)
-
 - ### [Install C++ on different OS](Install%20C%2B%2B%20on%20different%20OS/) 
 
-- ### [Introduction of OOPs in C++](Introduction%20of%20OOPs%20in%20C%2B%2B/)
+- ### [Fundamentals](./Fundamentals/)
 
-- ### [Classes & Objects Programming](Classes%20%26%20Objects%20Programming/)
+  - [Introduction of C/C++](./Fundamentals/Introduction%20of%20C/)
+  - [Programming components, Structure & Keyword](./Fundamentals/Programming%20components%2C%20Structure%20%26%20Keyword/)
+  - [Operators & Expressions](./Fundamentals/Operators%20%26%20Expressions/)
+  - [Input & Output Operations](./Fundamentals/Input%20%26%20Output%20Operations/)
+  - [Decision Making & Branching Structure](./Fundamentals/Decision%20Making%20%26%20Branching%20Structure/)
+  - [Looping Control Structure/Statements](./Fundamentals/Looping%20Control%20Structure%20or%20Statements/)
+  - [Break/Continue](./Fundamentals/Break%20or%20Continue/)
+  - [Static scoping & Dynamic Scoping](./Fundamentals/Static%20scoping%20%26%20Dynamic%20Scoping/)
+  - [Functions & Recursion and its types](./Fundamentals/Functions%20%26%20Recursion%20and%20its%20types/)
+  - [Arrays](./Fundamentals/Arrays/)
+  
+- ### [Strings and their functions](./Strings%20and%20their%20functions/)
+  
+- ### [Pointer and its types](./Pointer%20and%20its%20types/)
+  
+- ### [Procedural Structures](./Procedural%20Structures/)
 
-- ### [Constructor & Destructor](Constructor%20%26%20Destructor/) 
+- ### [Dynamic Memory allocation](./Dynamic%20Memory%20allocation/)
 
-- ### [Virtual Functions, Friend Functions & Friend Classes](Virtual%20Functions%2C%20Friend%20Functions%20%26%20Friend%20Classes/) 
+- ### [Structure padding, Union, Enum, & typedef](./Structure%20padding%2C%20Union%2C%20Enum%2C%20%26%20typedef/)
 
-- ### [Polymorphism / Overloading](Polymorphism%20or%20Overloading/)
+- ### [File Management / Handling](./File%20Management%20or%20Handling/)
 
-- ### [Inheritance and its types / Function overriding](Inheritance%20and%20its%20types%20or%20Function%20overriding/)
+- ### [Object Oriented Programming in C++](./OOPs%20in%20C++/)
 
-- ### [Encapsulation](Encapsulation/)
+  - [Introduction of OOPs in C++](./OOPs%20in%20C++/Introduction%20of%20OOPs%20in%20C%2B%2B/)
+  - [Classes & Objects Programming](./OOPs%20in%20C++/Classes%20%26%20Objects%20Programming/)
+  - [Constructor & Destructor](./OOPs%20in%20C++/Constructor%20%26%20Destructor/)
+  - [Data Abstraction](#)
+  - [Encapsulation](./OOPs%20in%20C++/Encapsulation/)
+  - [Polymorphism / Overloading](./OOPs%20in%20C++/Polymorphism%20or%20Overloading/)
+  - [Inheritance and its types / Function overriding](./OOPs%20in%20C++/Inheritance%20and%20its%20types%20or%20Function%20overriding/)
+  - [Virtual Functions, Friend Functions & Friend Classes](./OOPs%20in%20C++/Virtual%20Functions%2C%20Friend%20Functions%20%26%20Friend%20Classes/)
+  - [Operator Overloading](./OOPs%20in%20C++/Operator%20Overloading/)
 
-- ### [Operator Overloading](Operator%20Overloading/) 
+- ### [Advance C++ OOPs programming](./Advance%20C%2B%2B%20OOPs%20programming/)
 
-- ### [Advance C++ OOPs programming](Advance%20C%2B%2B%20OOPs%20programming/)
+- ### [Templates & Libraries](./Templates%20%26%20Libraries/)
 
-- ### [Templates & Libraries](Templates%20%26%20Libraries/)
+- ### [Exception Handling](./Exception%20Handling/)
 
-- ### [Exception Handling](Exception%20Handling/)
+- ### [File Handling & IO Management](./File%20Handling%20%26%20IO%20Management/)
 
-- ### [File Handling & IO Management](File%20Handling%20%26%20IO%20Management/)
-
-- ### [Number System i.e Binary, Octal, Decimal, HexaDecimal etc](Number%20System%20i.e%20Binary%2C%20Octal%2C%20Decimal%2C%20HexaDecimal%20etc/)
+- ### [Number System i.e Binary, Octal, Decimal, HexaDecimal etc](./Number%20System/)
 
 - ### [Data Structures](Data%20Structures/)
-
-  - ### [Standard Template Library (Vectors, Maps etc)](Data%20Structures/Standard%20Template%20Library%20(Vectors%2C%20Maps%20etc)/)
-
-  - ### [User Defined (Fenwick Tree, Segment Tree etc)](Data%20Structures/User%20Defined%20(Fenwick%20Tree%2C%20Segment%20Tree%20etc)/)
+  
+  - [Array](./Data%20Structures/Array)
+  - [Linked List](./Data%20Structures/LinkedList)
+  - [Stack](./Data%20Structures/Stack)
+  - [Queue](./Data%20Structures/Queue)
+  - [Binary Tree](./Data%20Structures/BinaryTree)
+  - [Binary Search Tree](./Data%20Structures/BinarySearchTree)
+  - [Heap](./Data%20Structures/Heap)
+  - [Hashing](./Data%20Structures/Hashing)
+  - [Graph](./Data%20Structures/Graph)
+  - [Matrix](./Data%20Structures/Matrix)
+  - [Trees](./Data%20Structures/Trees)
+  - [Standard Template Library](./Data%20Structures/Standard%20Template%20Library)
+    - [Vector](./Data%20Structures/Standard%20Template%20Library/Vector)
+    - [List](#)
+    - [Deque](./Data%20Structures/Standard%20Template%20Library/Deque/)
+    - [Queue](#)
+    - [Stack](./Data%20Structures/Standard%20Template%20Library/Stack/)
+    - [Priority Queue](./Data%20Structures/Standard%20Template%20Library/Priority_Queue/)
+  - [User Defined Data Structures](./UserDefinedDataStructures)
+    - [Segment Tree](#)
+    - [Fenwick Tree](#)
+    - [BTree](#)
+    - [Trie](#)
+    - [Suffix Tree](#)
   
 - ### [Operating System Concepts](Operating%20System%20Concepts%20/)
-
-  - ### [Scheduling Algorithms](#)
-
-  - ### [Memory Management](#)
-
-  - ### [Deadlock](#)
-
-  - ### [File Management](#)
+  - [Scheduling Algorithms](#)
+  - [Memory Management](#)
+  - [Deadlock](#)
+  - [File Management](#)
 
 - ### [Competitive Programming](Competitive%20Programming/)
 
 - ### [Algorithmic Approaches](Algorithmic%20Approaches/)
+  
+  - [Sorting Algorithm](./Algorithmic%20Approaches/Sorting)
+  - [Brute Force Algorithm](./Algorithmic%20Approaches/Brute%20Force)
+  - [Recursive Algorithm](./Algorithmic%20Approaches/Recursive)
+  - [Divide & Conquer Algorithm](./Algorithmic%20Approaches/Divide%20%26%20Conquer)
+  - [Greedy Algorithm](./Algorithmic%20Approaches/Greedy)
+  - [Dynamic Programming Algorithm](./Algorithmic%20Approaches/Dynamic)
+  - [Backtracking Algorithm](./Algorithmic%20Approaches/Backtracking)
+  - [Randomised Algorithm](./Algorithmic%20Approaches/Randomised)
+  - [Branch & Bound Algorithm](./Algorithmic%20Approaches/Branch%20%26%20Bound)
+  - [NP Complete Algorithm](./Algorithmic%20Approaches/NP%20Completeness)
 
 - ### [OpenGL using C++ (3D,2D graphics using C++)](OpenGL%20using%20C%2B%2B%20(3D%2C2D%20graphics%20using%20C%2B%2B)/)
